Add files via upload
this code is to be run on a client and a server. the client will have only clientimage_2.py file and the server will have trainer ,serverimage_6 and record_face.py.  first you need to train the LBPH classifier at the server and then make a client to server connection to check for the authentication of the client.